+++ custom_components/aprilaire_rs485/__init__.py
@@ -34,6 +34,7 @@ PLATFORMS: list[Platform] = [
     Platform.BUTTON,
     Platform.CLIMATE,
     Platform.HUMIDIFIER,
+    Platform.SELECT,
     Platform.SENSOR,
     Platform.TEXT,
 ]